+++ src/LIN3S/WPFoundation/PostTypes/Fields/FieldConnector.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace LIN3S\WPFoundation\PostTypes\Fields;
+
+
+interface FieldConnector
+{
+    /**
+     * @return array
+     */
+    public function connector();
+}